# feat: parallel gzip compression/decompressio + adjustments to ReadLine