--- a/fileio/byteio.go
+++ b/fileio/byteio.go
@@ -11,8 +11,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/klauspost/pgzip"
 	"github.com/vertgenlab/gonomics/exception"
+	"github.com/klauspost/pgzip"
+	
 )
 
 const (
@@ -31,74 +32,85 @@ const (
 // The drawback is that ByteReader is not as easy to use as EasyReader
 // and should generally be reserved for performance intensive tasks.
 type ByteReader struct {
-	*bufio.Reader
-	File         *os.File
-	internalGzip *pgzip.Reader // Use pgzip.Reader here
-	line         []byte
-	Buffer       *bytes.Buffer
+    *bufio.Reader
+    File         *os.File
+    internalGzip *pgzip.Reader // Use pgzip.Reader here
+    line         []byte
+    Buffer       *bytes.Buffer
 }
 
 // Read reads data into p and is a method required to implement the io.Reader interface.
 // It returns the number of bytes read into p.
 func (reader *ByteReader) Read(b []byte) (n int, err error) {
-	if reader.internalGzip != nil {
-		return reader.internalGzip.Read(b)
-	}
-	return reader.Reader.Read(b)
+    if reader.internalGzip != nil {
+        return reader.internalGzip.Read(b)
+    }
+    return reader.Reader.Read(b)
 }
+
 
 // NewByteReader will process a given file and performs error handling if an error occurs.
 // ByteReader will process gzipped files accordingly by performing a check on the suffix
 // of the provided file.
 func NewByteReader(filename string) *ByteReader {
-	file := MustOpen(filename)
+    file, err := os.Open(filename)
+    exception.PanicOnErr(err)
+	
+    reader := &ByteReader{
+        File:   file,
+        Buffer: &bytes.Buffer{},
+    }
 
-	reader := &ByteReader{
-		File:   file,
-		Buffer: &bytes.Buffer{},
-	}
+    if strings.HasSuffix(filename, ".gz") {
+        // Initialize pgzip.Reader for gzip files
+        gzReader, err := pgzip.NewReader(file)
+        exception.PanicOnErr(err)
 
-	if strings.HasSuffix(filename, ".gz") {
-		var err error
-		reader.internalGzip, err = pgzip.NewReader(file)
-		exception.PanicOnErr(err)
-
-		reader.Reader = bufio.NewReader(reader.internalGzip)
-	} else {
-		reader.Reader = bufio.NewReader(file)
-	}
-	return reader
+        reader.internalGzip = gzReader
+        reader.Reader = bufio.NewReader(gzReader)
+    } else {
+        reader.Reader = bufio.NewReader(file)
+    }
+    return reader
 }
+
+
 
 // ReadLine will return a bytes.Buffer pointing to the internal slice of bytes. Provided this function is called within a loop,
 // the function will read one line at a time, and return bool to continue reading. Important to note the buffer return points to
 // the internal slice belonging to the reader, meaning the slice will be overridden if the data is not copied.
 func ReadLine(reader *ByteReader) (*bytes.Buffer, bool) {
-	reader.Buffer.Reset() // Reset buffer for new line reading
-	var err error
-	for reader.line, err = reader.ReadSlice('\n'); err != nil || err != bufio.ErrBufferFull; reader.line, err = reader.ReadSlice('\n') {
-		if err == io.EOF {
-			return reader.Buffer, true // End of file reached
-		}
-		// Write the read part to the buffer, handling both partial and complete lines
-		if len(reader.line) > 0 && reader.line[len(reader.line)-1] == '\n' {
-			// Check for carriage return before newline and adjust accordingly
-			if len(reader.line) > 1 && reader.line[len(reader.line)-2] == '\r' {
-				_, err = reader.Buffer.Write(reader.line[:len(reader.line)-2])
-			} else {
-				_, err = reader.Buffer.Write(reader.line[:len(reader.line)-1])
-			}
-			exception.PanicOnErr(err)
-			break // Complete line read, exit loop
-		} else {
-			_, err = reader.Buffer.Write(reader.line)
-			exception.PanicOnErr(err)
-			if err == bufio.ErrBufferFull {
-				continue // Buffer was full, continue reading the line
-			}
-		}
-	}
-	return reader.Buffer, false
+    reader.Buffer.Reset() // Reset buffer for new line reading
+    var err error
+    for {
+        reader.line, err = reader.ReadSlice('\n')
+        if err != nil && err != bufio.ErrBufferFull {
+            if err == io.EOF {
+                return reader.Buffer, true // End of file reached
+            }
+            exception.PanicOnErr(err) // Handle unexpected errors
+        }
+        // Write the read part to the buffer, handling both partial and complete lines
+        if len(reader.line) > 0 && reader.line[len(reader.line)-1] == '\n' {
+            // Check for carriage return before newline and adjust accordingly
+            if len(reader.line) > 1 && reader.line[len(reader.line)-2] == '\r' {
+                _, err = reader.Buffer.Write(reader.line[:len(reader.line)-2])
+            } else {
+                _, err = reader.Buffer.Write(reader.line[:len(reader.line)-1])
+            }
+            exception.PanicOnErr(err)
+            if err == nil {
+                break // Complete line read, exit loop
+            }
+        } else {
+            _, err = reader.Buffer.Write(reader.line)
+            exception.PanicOnErr(err)
+            if err == bufio.ErrBufferFull {
+                continue // Buffer was full, continue reading the line
+            }
+        }
+    }
+    return reader.Buffer, false
 }
 
 // CatchErrThrowEOF will silently handles and throws the EOF error and will log and exit any other errors.