--- a/fileio/byteio.go
+++ b/fileio/byteio.go
@@ -121,7 +121,7 @@ func (br *ByteReader) Close() error {
 		return errors.New("no file found")
 	}
 
-	switch {
+	switch { // Handle error returns. Priority is gzErr > fileErr
 	case gzErr != nil:
 		return gzErr