# Description
Pgzip to replace golang's native gzip could have huge performance improvements as well, so we can also look to implement this into **EasyReader{}** as well
**Important changes:**
https://github.com/vertgenlab/gonomics/blob/5d13023f4faaf53d7b43e2909e55623f16b458c4/fileio/byteio.go#L23-L29
- Additionally, I added a few fixes to known bugs for a while, mostly how to deal with long times and rewrote the godoc strings to be concise and clearer.
- Most changes are reduces copies and check due to previous implementation, but the most important bug fix is:
https://github.com/vertgenlab/gonomics/blob/5d13023f4faaf53d7b43e2909e55623f16b458c4/fileio/byteio.go#L32-L37

#### The native  gzip is know to be not as performant particularly in Golang. here are some of the initial benchmarks:
## Relevant Links
https://github.com/klauspost/pgzip

### Before:
```tsv
BenchmarkByteioReader-12             162          73653100 ns/op            4488 B/op          8 allocs/op
BenchmarkByteioReaderGz-12            19         628690623 ns/op          326833 B/op       4567 allocs/op
```

### After:
```tsv
BenchmarkByteioReader-12             204          69317792 ns/op            4488 B/op          8 allocs/op
BenchmarkByteioReaderGz-12            31         452670788 ns/op         4256291 B/op        225 allocs/op

```

### Comparison to rest:
```tsv
BenchmarkByteioReader-12             204          70612963 ns/op            4488 B/op          8 allocs/op
BenchmarkByteioReaderGz-12            31         461232750 ns/op         4256613 B/op        226 allocs/op
BenchmarkEasyReaderReg-12            100         122885649 ns/op        217607149 B/op   3400036 allocs/op
BenchmarkEasyReaderGz-12              20         671629554 ns/op        217928840 B/op   3404586 allocs/op
BenchmarkRead-12                     100         123999460 ns/op
BenchmarkReadGz-12                    20         677874577 ns/op
```

### Testing
<!-- if relevant, document how you tested this code, and how someone else might also test it -->
```bash
go test -bench=BenchmarkByteioReader -benchtime=12s
```

### Checklist before requesting a review

- [x] I performed a self-review of my code
- [x] If it this a core feature, I added thorough tests
- [x] The command `go fmt` or `make clean` was used on all files included

<!-- ### Screenshots & Media
if relevant, add an screenshots, images or recordings -->

<!-- ### Edge cases / Breaking Changes / Known Issues
if relevant, document any edge cases, known issues, etc -->